Update README.md
Logistics

For the course project, you will implement a research idea related to the course material. The purpose of the nal project is to give you some experience working on a piece of original research and writing up your results in a paper style format. You are expected to describe your research idea/application clearly in the project proposal, relate to existing work. You will document the project progress in the nal report.

You must form a group of two or three to complete the project. Your report must clearly list the contributions of each team member. Once your group is formed, please sign up your group through Quercus. Instructions of sign-up can be found here: https://qstudents. utoronto.ca/group-tool-the-student-side-of-things/

Academic Integrity

Any work or ideas that are not your own must be appropriately acknowledged with a citation in the nal report. This include any code used for the experiments and text in the report itself. It is important to note that plagiarism and other forms of academic misconduct are not tolerated in this course. The University of Toronto’s Code of Behaviour on Academic Matters outlines the behaviours that constitute academic misconduct, the processes for addressing academic o enses, and the penalties that may be imposed. You are expected to be familiar with the contents of this document.

Large Language Model Policy

We will adopt a similar guideline to the International Conference on Machine Learning 2023 policy about the usage of Large Language Models (LLMs), such as ChatGPT, for the nal report. The LLM policy prohibits text produced entirely by LLMs (i.e., \generated”). This does not prohibit authors from using LLMs for editing or polishing author-written text.

The LLM policy is largely predicated on the principle of being conservative with respect to guarding against potential issues of using LLMs, including plagiarism.

Writing format